10th venki2

+++ src/Models/ChatRoom_User.java
@@ -6,10 +6,10 @@ import org.json.JSONObject;
 public class ChatRoom_User {
 
 	private static int id;
-	private static int crid;
-	private static int usrid;
-	private static int Cur_Mak;
-	private static int Cur_Ekey;
+	private static long crid;
+	private static long usrid;
+	private static long Cur_Mak;
+	private static long Cur_Ekey;
 	
 	public static final String CRU_CRID="current user chat room id";
 	public static final String CRU_CUR_MAK = "current user current MAK";
@@ -20,7 +20,7 @@ public class ChatRoom_User {
 	      
      }
      // constructor
-     public ChatRoom_User(int id,int crid, int usrid, int Cur_Mak, int Cur_Ekey){
+     public ChatRoom_User(int id,long crid, long usrid, long Cur_Mak, long Cur_Ekey){
          this.id      = id;
          this.crid=crid;
          this.Cur_Mak    = Cur_Mak;
@@ -37,36 +37,36 @@ public class ChatRoom_User {
 	public void setID(int id){
         this.id = id;
     }
-	public int getCur_Mak(){
+	public long getCur_Mak(){
         return this.Cur_Mak;
     }
 
-    public void setCur_Mak(int Cur_Mak){
+    public void setCur_Mak(long Cur_Mak){
         this.Cur_Mak = Cur_Mak;
     }
   
    
-    public void setCur_Ekey(int Cur_Ekey){
+    public void setCur_Ekey(long Cur_Ekey){
         this.Cur_Ekey = Cur_Ekey;
     }
     
-    public int getCur_Ekey(){
+    public long getCur_Ekey(){
         return this.Cur_Ekey;
     }
     
-    public void setusrid(int usrid){
+    public void setusrid(long usrid){
         this.usrid = usrid;
     }
     
-    public int getusrid(){
+    public long getusrid(){
         return this.usrid;
     }  
 	
-    public void setcrid(int crid){
+    public void setcrid(long crid){
         this.crid = crid;
     }
     
-    public int getcrid(){
+    public long getcrid(){
         return this.crid;
     }
